refactor(harness): extract shared workspace watch ownership
Move the canonical-root watcher and shared lease broker behind a graph-neutral internal module while preserving session and legacy graph adapters.

Closes: SAP-3088
