+++ .changeset/calm-watches-share.md
@@ -1,0 +1,5 @@
+---
+"@sapiom/harness": patch
+---
+
+Move shared workspace watch ownership behind a graph-neutral internal module without changing watcher behavior.
